[script] Main sync logic
The script is now able to detect the presence of dev and dotfiles
target in the user home directory.

The test can be run in normal and test mode.
While running the script in test, setting to TRUE the proper
parameter within the script, a test folder is created and target
directories change.

The next step could be to accept parameters in input
- prompt an help message
- guide the setup of work and personal envs
- change the target directories
- flag the test mode via parameter.

We could also create a test.sh script to test the several working
scenarios